+++ backend/src/middlewares/index.ts
@@ -1,2 +1,3 @@
-export * from './errorMiddleware';
-export * from './loggerMiddleware';
+export * from './error-handler';
+export * from './logger';
+export * from './validation';
