+++ backend/src/middlewares/logger.ts
@@ -9,7 +9,9 @@ export const loggerMiddleware = (req: Request, res: Response, next: NextFunction
     const endTime = Date.now();
     const elapsedTime = endTime - startTime;
 
-    console.log(`[${new Date().toISOString()}] ${req.method} ${req.url} ${res.statusCode} - ${elapsedTime}ms`);
+    console.log(
+      `[${new Date().toISOString()}] ${req.method} ${req.url} ${res.statusCode} - ${elapsedTime}ms`
+    );
 
     return originalSend.apply(res, args);
   };
